+++ tiledesk-client/mqtt-route.js
@@ -1,3 +1,6 @@
+/**
+ * Route to add on the server to test for health of MQTT messaging
+ */
 const express = require('express');
 const router = express.Router();
 const axios = require('axios');